All cave_hashmp functions implimented. Missing tests and hash fn's
There needs to be quite a bit more documentation and very thrurogh testing. Bound to be plenty of bugs. In particular I have a hunch that the CaveHashMap.count is not always being updated when it should be. Each function will get thoroughly run down as I write tests for them.